finish the homepage header autoscroll, and merged navigation drawer

+++ kitchen_list/src/main/java/edu/scu/qjiang/homekitchen/entities/DishItem.java
@@ -14,7 +14,7 @@ public class DishItem
   private java.util.Date updated;
   private String description;
   private String name;
-  private Integer price;
+  private Double price;
   private String picture;
   private String objectId;
   private java.util.Date created;
@@ -58,12 +58,12 @@ public class DishItem
     this.name = name;
   }
 
-  public Integer getPrice()
+  public Double getPrice()
   {
     return price;
   }
 
-  public void setPrice( Integer price )
+  public void setPrice( Double price )
   {
     this.price = price;
   }
